requirements.txt
PySide6==6.7.2
pyserial==3.5
pygments==2.18.0
pyqtgraph==0.13.7
GitPython==3.1.43
requests==2.32.3
jedi==0.19.1
